# Disable stack-protector for alloca() fns