Doesn't really seem to move the needle much (I verified that it remove the calls to stkchk)...

before:
```
Benchmark                                            Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------
BM_for_each_contiguous_slice/64/16/1              41.0 ns         41.0 ns     65728123 bytes_per_second=45.0949Gi/s
BM_for_each_contiguous_slice/64/4/4               41.1 ns         41.1 ns     70735311 bytes_per_second=44.9251Gi/s
BM_for_each_contiguous_slice_multi/64/16/1        62.4 ns         62.4 ns     45553978 bytes_per_second=59.2437Gi/s
BM_for_each_contiguous_slice_multi/64/4/4         60.2 ns         60.2 ns     45854848 bytes_per_second=61.3731Gi/s
BM_for_each_contiguous_slice_multi3/64/16/1        595 ns          595 ns      4609380 bytes_per_second=9.32335Gi/s
BM_for_each_contiguous_slice_multi3/64/4/4         595 ns          595 ns      4601328 bytes_per_second=9.32106Gi/s
```

after:
```
------------------------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------
BM_for_each_contiguous_slice/64/16/1              38.8 ns         38.8 ns     72082318 bytes_per_second=47.5959Gi/s
BM_for_each_contiguous_slice/64/4/4               40.2 ns         40.2 ns     71936728 bytes_per_second=45.9804Gi/s
BM_for_each_contiguous_slice_multi/64/16/1        61.9 ns         61.9 ns     45927247 bytes_per_second=59.6876Gi/s
BM_for_each_contiguous_slice_multi/64/4/4         63.0 ns         62.9 ns     43511005 bytes_per_second=58.7059Gi/s
BM_for_each_contiguous_slice_multi3/64/16/1        599 ns          599 ns      4682256 bytes_per_second=9.25885Gi/s
BM_for_each_contiguous_slice_multi3/64/4/4         608 ns          608 ns      4635358 bytes_per_second=9.12313Gi/s
```